@@ -4,13 +4,12 @@ import {
   Input,
   ElementRef,
   AfterViewInit,
-} from '@angular/core';
-import KTLayoutExamples from '../../../../../../assets/js/layout/extended/examples';
+} from "@angular/core";
 
 @Component({
-  selector: 'app-code-preview',
-  templateUrl: './code-preview.component.html',
-  styleUrls: ['./code-preview.component.scss'],
+  selector: "app-code-preview",
+  templateUrl: "./code-preview.component.html",
+  styleUrls: ["./code-preview.component.scss"],
 })
 export class CodePreviewComponent implements OnInit, AfterViewInit {
   // Public properties
@@ -19,10 +18,5 @@ export class CodePreviewComponent implements OnInit, AfterViewInit {
 
   ngOnInit(): void {}
 
-  ngAfterViewInit() {
-    const elements = this.el.nativeElement.querySelectorAll(
-      '.example.example-compact'
-    );
-    KTLayoutExamples.init(elements);
-  }
+  ngAfterViewInit() {}
 }
