@@ -1,23 +1,15 @@
-import { Component, OnInit, AfterViewInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 import { LayoutService } from '../../../../_metronic/core';
 import { AuthService } from '../../../../modules/auth/_services/auth.service';
-import { UserModel } from '../../../../modules/auth/_models/user.model';
-import KTLayoutQuickSearch from '../../../../../assets/js/layout/extended/quick-search';
-import KTLayoutQuickNotifications from '../../../../../assets/js/layout/extended/quick-notifications';
-import KTLayoutQuickActions from '../../../../../assets/js/layout/extended/quick-actions';
-import KTLayoutQuickCartPanel from '../../../../../assets/js/layout/extended/quick-cart';
-import KTLayoutQuickPanel from '../../../../../assets/js/layout/extended/quick-panel';
-import KTLayoutQuickUser from '../../../../../assets/js/layout/extended/quick-user';
-import KTLayoutHeaderTopbar from '../../../../../assets/js/layout/base/header-topbar';
-import { KTUtil } from '../../../../../assets/js/components/util';
+import { UserModel } from '../../../../modules/auth/_models/user.model'; 
 
 @Component({
   selector: 'app-topbar',
   templateUrl: './topbar.component.html',
   styleUrls: ['./topbar.component.scss'],
 })
-export class TopbarComponent implements OnInit, AfterViewInit {
+export class TopbarComponent implements OnInit {
   user$: Observable<UserModel>;
   // tobbar extras
   extraSearchDisplay: boolean;
@@ -64,48 +56,5 @@ export class TopbarComponent implements OnInit, AfterViewInit {
       'extras.quickPanel.display'
     );
   }
-
-  ngAfterViewInit(): void {
-    KTUtil.ready(() => {
-      // Called after ngAfterContentInit when the component's view has been initialized. Applies to components only.
-      // Add 'implements AfterViewInit' to the class.
-      if (this.extraSearchDisplay && this.extrasSearchLayout === 'offcanvas') {
-        KTLayoutQuickSearch.init('kt_quick_search');
-      }
-
-      if (
-        this.extrasNotificationsDisplay &&
-        this.extrasNotificationsLayout === 'offcanvas'
-      ) {
-        // Init Quick Notifications Offcanvas Panel
-        KTLayoutQuickNotifications.init('kt_quick_notifications');
-      }
-
-      if (
-        this.extrasQuickActionsDisplay &&
-        this.extrasQuickActionsLayout === 'offcanvas'
-      ) {
-        // Init Quick Actions Offcanvas Panel
-        KTLayoutQuickActions.init('kt_quick_actions');
-      }
-
-      if (this.extrasCartDisplay && this.extrasCartLayout === 'offcanvas') {
-        // Init Quick Cart Panel
-        KTLayoutQuickCartPanel.init('kt_quick_cart');
-      }
-
-      if (this.extrasQuickPanelDisplay) {
-        // Init Quick Offcanvas Panel
-        KTLayoutQuickPanel.init('kt_quick_panel');
-      }
-
-      if (this.extrasUserDisplay && this.extrasUserLayout === 'offcanvas') {
-        // Init Quick User Panel
-        KTLayoutQuickUser.init('kt_quick_user');
-      }
-
-      // Init Header Topbar For Mobile Mode
-      KTLayoutHeaderTopbar.init('kt_header_mobile_topbar_toggle');
-    });
-  }
+ 
 }
