@@ -1,18 +1,11 @@
-import { Component, OnInit, AfterViewInit } from '@angular/core';
-import { KTUtil } from '../../../../../assets/js/components/util';
-import KTLayoutAsideToggle from '../../../../../assets/js/layout/base/aside-toggle';
-import KTLayoutStickyCard from '../../../../../assets/js/layout/base/sticky-card';
-import KTLayoutStretchedCard from '../../../../../assets/js/layout/base/stretched-card';
-import { LayoutService } from '../../../../_metronic/core';
-import KTLayoutBrand from '../../../../../assets/js/layout/base/brand';
-import KTLayoutAside from '../../../../../assets/js/layout/base/aside';
-import KTLayoutAsideMenu from '../../../../../assets/js/layout/base/aside-menu';
+import { Component, OnInit } from '@angular/core'; 
+import { LayoutService } from '../../../../_metronic/core'; 
 
 @Component({
   selector: 'app-scripts-init',
   templateUrl: './scripts-init.component.html',
 })
-export class ScriptsInitComponent implements OnInit, AfterViewInit {
+export class ScriptsInitComponent implements OnInit {
   asideSelfMinimizeToggle = false;
 
   constructor(private layout: LayoutService) { }
@@ -22,25 +15,5 @@ export class ScriptsInitComponent implements OnInit, AfterViewInit {
       'aside.self.minimize.toggle'
     );
   }
-
-  ngAfterViewInit() {
-    KTUtil.ready(() => {
-      // Init Brand Panel For Logo
-      KTLayoutBrand.init('kt_brand');
-      // Init Aside
-      KTLayoutAside.init('kt_aside');
-      // Init Aside Menu
-      KTLayoutAsideMenu.init('kt_aside_menu');
-
-      if (this.asideSelfMinimizeToggle) {
-        // Init Aside Menu Toggle
-        KTLayoutAsideToggle.init('kt_aside_toggle');
-      }
-
-      // Init Sticky Card
-      KTLayoutStickyCard.init('kt_page_sticky_card');
-      // Init Stretched Card
-      KTLayoutStretchedCard.init('kt_page_stretched_card');
-    });
-  }
+ 
 }
